requirements.txt
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.15.1
torch==2.2.2
transformers==4.44.2
sentence-transformers==5.1.1
faiss-cpu==1.7.4
fastapi
apscheduler